# Conflicts:
#   src/main/java/com/trade_accounting/controllers/rest/EmployeeRestController.java
#   src/main/java/com/trade_accounting/services/impl/EmployeeServiceImpl.java
#   src/main/java/com/trade_accounting/services/interfaces/EmployeeService.java
#   src/test/java/com/trade_accounting/services/impl/EmployeeServiceImplTest.java

+++ src/main/java/com/trade_accounting/services/interfaces/fias/CityService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface CityService {
     List<CityDto> getAll();
 
-    CityDto getById(Integer id);
+    CityDto getById(Long id);
 
     CityDto create(CityDto cityDto);
 
     CityDto update(CityDto cityDto);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
